Merge dev into master

+++ dsutil/docker/docker.py
@@ -166,15 +166,22 @@ class DockerImage:
 
     def _copy_ssh(self, copy_ssh_to: str):
         if copy_ssh_to:
+            ssh_src = Path.home() / ".ssh"
+            if not ssh_src.is_dir():
+                logger.warning("~/.ssh does NOT exists!")
+                return
             ssh_dst = self.path / copy_ssh_to
             try:
                 shutil.rmtree(ssh_dst)
             except FileNotFoundError:
                 pass
-            for path in (Path.home() / ".ssh").iterdir():
+            ssh_dst.mkdir()
+            ssh_dst.chmod(0o700)
+            for path in ssh_src.iterdir():
                 if path.is_file():
                     shutil.copy2(path, ssh_dst)
                     (ssh_dst / path.name).chmod(0o600)
+            logger.info("~/.ssh has been copied to {}", ssh_dst)
 
     def build(
         self,
